ユニコの脳みそ/00/Legacy_Tools/インスタコメント返信投稿予約/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
apscheduler==3.10.4
requests==2.31.0
gspread==5.12.0
google-auth==2.25.0
python-dotenv==1.0.0


